define source and query with sqexpr
