+++ packages/malloy/src/lang/ast/index.ts
@@ -27,6 +27,7 @@ export * from './elements/define-query';
 export * from './elements/source';
 export * from './elements/refined-source';
 export * from './elements/source-query';
+export * from './elements/source-query-nodes';
 export * from './source-properties/field-list-edit';
 export * from './source-properties/primary-key';
 export * from './source-properties/renames';
