+++ packages/malloy/src/lang/test/locations.spec.ts
@@ -371,13 +371,13 @@ describe('source references', () => {
   });
 
   test('reference to query in from', () => {
-    const source = markSource`
+    const source = model`
       query: ${'q is a -> { project: * }'}
       source: na is from(-> ${'q'})
     `;
-    const m = new TestTranslator(source.code);
-    expect(m).toTranslate();
-    expect(m.referenceAt(pos(source.locations[1]))).toMatchObject({
+    expect(source).toTranslate();
+    const mt = source.translator;
+    expect(mt.referenceAt(pos(source.locations[1]))).toMatchObject({
       location: source.locations[1],
       type: 'queryReference',
       text: 'q',
